requirements.txt
streamlit
chromadb
sentence-transformers
groq
python-dotenv
PyPDF2